Move constants file and add models filepath
